requirements.txt
Flask
deepface
Pillow
numpy
